+++ app/components/UI/banner/Button.js
@@ -1,0 +1,12 @@
+export default function Button() {
+  return (
+    <div>
+      <a
+        href="#"
+        className="inline-block rounded-md border border-transparent bg-indigo-600 px-8 py-3 text-center font-medium text-white hover:bg-indigo-700"
+      >
+        Pedido de Roteiro Gratuito
+      </a>
+    </div>
+  );
+}
